executare algoritm pas cu pas. Mai e de facut citirea

+++ fereastra.cpp
@@ -10,7 +10,7 @@ void creareFereastra()
 {
 	VideoMode desktop = VideoMode::getDesktopMode();
 	RenderWindow fereastraAplicatie(desktop, "Interschem", Style::Default); //de schimbat cu Fullscreen dupa
-	const View view(FloatRect(0, 0, desktop.width, desktop.height));
+	const View view(FloatRect(0, 0, static_cast<float>(desktop.width), static_cast<float>(desktop.height)));
 	fereastraAplicatie.setView(view);
 	fereastraAplicatie.setFramerateLimit(60);//limita de fps ca sa nu bubuie laptopul :')
 	fereastraAplicatie.setKeyRepeatEnabled(false);
